+++ src/bot/commands/puzzle.rs
